requirements.txt
django
djangorestframework
requests
beautifulsoup4
redis
django-redis
aiohttp
pandas